Make the branch PHPStan-clean and fix a null-contest crash in the jury menu
Closing the second gap: I had never run PHPStan or the unit tests, though my own
phase 1 said to record both as a baseline. Measured against pristine 9.0.1 under
the same PHP version, since the fork's numbers are meaningless on their own.

PHPStan: pristine 9.0.1 reports 0 errors, the branch reported 32. Now 0.
(An earlier count of 222 was an artefact of running under PHP 8.5 rather than
8.3 - most of those were spurious "attribute is not repeatable" reports.)

Real defects among them, all inherited from 8.3.2:

- API ClarificationController called parent::__construct() with no arguments,
  but 9.0's AbstractRestController requires four services. Every clarification
  API route would have died with a TypeError. It also referenced an undefined
  $newClarification when notifying Discord - the variable in scope is
  $clarification.
- The team editor carried #[Security(...)] from Sensio's FrameworkExtraBundle,
  which no longer exists; replaced with upstream's
  #[IsGranted(new Expression(...))] idiom.
- ProblemTestcaseGroupController tested empty($lockedContest) - singular, never
  defined - so the locked-contest guard never applied and an admin could edit
  testcase groups of a locked contest. The variable is $lockedContests.
- TeamScore::$numPoints had been widened to float, which leaked into
  ResultRow::$numSolved and Score::$numSolved. Both are int because the CCS spec
  defines num_solved as an integer, and rankcache.points_* is an int column; the
  fractional score lives in scorecache.points_* and the sort key. Reverted.

The rest were annotation debt in fork-only files: missing iterable value types
and return types, Collection generics, and @var docblocks that claimed
find() results were non-null, which made PHPStan read the necessary null guards
as dead code.

Unit tests, same environment both sides (the suite needs a full nginx install
this box does not have, so the ~1000 errors are environmental and present in
both):

    pristine 9.0.1   errors 1057   failures  37
    branch (before)  errors  997   failures 521
    branch (after)   errors 1031   failures  45

498 of the 521 failures had one cause: menu_change_contest.html.twig line 5
called contest.getState() unguarded, so every jury page 500'd whenever there was
no current contest - the same null-contest pattern already fixed twice in this
migration. Upstream guards the same variable with |default('no contest') one
line below. Guarded; 476 failures went away.

The residual 45 vs 37 is fork UI divergence in content assertions.

+++ webapp/src/Utils/Scoreboard/TeamScore.php
@@ -7,7 +7,7 @@ use App\Entity\Team;
 
 class TeamScore
 {
-    public float $numPoints = 0;
+    public int $numPoints = 0;
 
     public int $rank = 0;
     public int $totalTime;
